DEPRECATED: Coerce to tibble. Enable preserving row names when coercing matrix and time-series-like objects with row names.

Coerce to tibble. Enable preserving row names when coercing matrix and time-series-like objects with row names.

DEPRECATED: Use timetk::tk_tbl() instead.

as_tibble(x, preserve_row_names = FALSE, ...)

Arguments

x
A list, matrix, xts, zoo, timeSeries, etc object.

preserve_row_names
Used during coercion from matrix, xts, zoo, timeSeries, etc objects that have row names. When TRUE, creates a row.names column with names of rows as character class.

...
Additional parameters passed to the appropriate timetk::tk_tbl() function.

Value

Returns a tibble object.

Details

as_tibble is a wrapper for tibble::as_tibble that includes a preserve_row_names argument. The function is designed to coerce xts, zoo, timeSeries, ts, and irts objects that are used frequently in quantitative financial analysis. When preserve_row_names = TRUE is specified, a new column, row.names, is created during object coercion as a character class.

See also

tk_xts() - Coercion to xts, replaces tidyquant::as_xts()

tk_tbl() - Coercion to tbl, replaces tidyquant::as_tibble()

as_xts() - Deprecated, use tk_xts() instead.

Examples

# Load libraries
library(tidyquant)

# Matrix coercion to tibble
m <- matrix(rnorm(50), ncol = 5)
colnames(m) <- c("a", "b", "c", "d", "e")
rownames(m) <- letters[1:nrow(m)]
m_tbl <- as_tibble(m, preserve_row_names = TRUE)

#> Warning: The `tidyquant::as_tibble()` function is deprecated. Please use `timetk::tk_tbl()` instead.

# xts coercion to tibble
quantmod::getSymbols("AAPL", auto.assign = FALSE) %>%
    as_tibble(preserve_row_names = TRUE)

#> Warning: The `tidyquant::as_tibble()` function is deprecated. Please use `timetk::tk_tbl()` instead.

#> # A tibble: 2,675 x 7
#>         index AAPL.Open AAPL.High AAPL.Low AAPL.Close AAPL.Volume AAPL.Adjusted
#>        <date>     <dbl>     <dbl>    <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2007-01-03    13.702    13.748   13.005   11.97143   309579900      10.77017
#>  2 2007-01-04    13.346    13.648   13.310   12.23714   211815100      11.00922
#>  3 2007-01-05    13.619    13.688   13.402   12.15000   208685400      10.93082
#>  4 2007-01-08    13.650    13.740   13.542   12.21000   199276700      10.98480
#>  5 2007-01-09    13.727    14.764   13.521   13.22429   837324600      11.89731
#>  6 2007-01-10    15.045    15.530   14.839   13.85714   738220000      12.46666
#>  7 2007-01-11    15.234    15.368   15.101   13.68572   360063200      12.31244
#>  8 2007-01-12    15.020    15.095   14.804   13.51714   328172600      12.16078
#>  9 2007-01-16    15.193    15.442   15.157   13.87143   311019100      12.47951
#> 10 2007-01-17    15.492    15.498   15.057   13.56428   411565000      12.20319
#> # ... with 2,665 more rows