DEPRECATED: Coerce objects to xts, designed to work with tibble and data.frame objects

Coerce objects to xts, designed to work with tibble and data.frame objects

DEPRECATED: Use timetk::tk_xts() instead.

as_xts(x, date_col = NULL, ...)

as_xts_(x, date_col = NULL, ...)

Arguments

x
A data.frame (with date column), matrix, xts, zoo, timeSeries, etc object.

date_col
Required for objects that inherit the data.frame class. Must specify a date column that is of the date class. Unused for non-data.frame objects.

...
Additional parameters passed to timetk::tk_xts().

Value

Returns a xts object.

Details

as_xts is a wrapper for xts::as.xts that includes a date_col argument. When date_col is specified, the date column is used as row names during coercion to xts class. The date column must be in a date format (i.e. date class). as_xts_ evaluates the date_col using Non-Standard Evaluation (NSE). See vignette("nse") for details.

It is possible to coerce non-data.frame-like objects including zoo, timeSeries, ts, and irts objects. There is no need to specify the date_col argument.

See also

tk_xts() - Coercion to xts, replaces tidyquant::as_xts()

tk_tbl() - Coercion to tbl, replaces tidyquant::as_tibble()

as_tibble() - Deprecated, use tk_tbl() instead.

Examples

# Load libraries
library(tidyquant)

tq_get("AAPL", get = "stock.prices") %>%
    as_xts(date_col = date) # Deprecated: Use tk_xts()

#> Warning: The `as_xts()` function is deprecated. Please use `timetk::tk_xts()` instead.

#>               open    high     low     close    volume  adjusted
#> 2007-01-03  13.702  13.748  13.005  11.97143 309579900  10.77017
#> 2007-01-04  13.346  13.648  13.310  12.23714 211815100  11.00922
#> 2007-01-05  13.619  13.688  13.402  12.15000 208685400  10.93082
#> 2007-01-08  13.650  13.740  13.542  12.21000 199276700  10.98480
#> 2007-01-09  13.727  14.764  13.521  13.22429 837324600  11.89731
#> 2007-01-10  15.045  15.530  14.839  13.85714 738220000  12.46666
#> 2007-01-11  15.234  15.368  15.101  13.68572 360063200  12.31244
#> 2007-01-12  15.020  15.095  14.804  13.51714 328172600  12.16078
#> 2007-01-16  15.193  15.442  15.157  13.87143 311019100  12.47951
#> 2007-01-17  15.492  15.498  15.057  13.56428 411565000  12.20319
#> 2007-01-18  14.625  14.626  14.140  12.72429 591151400  11.44748
#> 2007-01-19  14.074  14.236  13.993  12.64286 341118400  11.37422
#> 2007-01-22  14.155  14.158  13.600  12.39857 363506500  11.15445
#> 2007-01-23  13.613  13.896  13.578  12.24286 301856100  11.01436
#> 2007-01-24  13.764  13.839  13.669  12.38571 231953400  11.14288
#> 2007-01-25  13.832  14.053  13.661  12.32143 226493400  11.08505
#> 2007-01-26  13.832  13.874  13.496  12.19714 246718500  10.97323
#> 2007-01-29  13.704  13.759  13.581  12.27714 225416100  11.04520
#> 2007-01-30  13.724  13.734  13.537  12.22143 144492600  10.99508
#> 2007-01-31  13.475  13.656  13.394  12.24714 214017300  11.01821
#> 2007-02-01  13.693  13.699  13.456  12.10571 166085500  10.89098
#> 2007-02-02  13.357  13.537  13.291  12.10714 155382500  10.89227
#> 2007-02-05  13.386  13.534  13.329  11.99143 144713100  10.78816
#> 2007-02-06  13.410  13.413  13.157  12.02143 216098400  10.81515
#> 2007-02-07  13.415  13.716  13.267  12.30714 266706300  11.07219
#> 2007-02-08  13.566  13.737  13.562  12.31143 169757700  11.07605
#> 2007-02-09  13.637  13.688  13.213  11.89571 215135200  10.70205
#> 2007-02-12  13.407  13.526  13.280  12.12571 181017900  10.90897
#> 2007-02-13  13.523  13.543  13.386  12.10000 145246500  10.88584
#> 2007-02-14  13.438  13.599  13.429  12.18572 126995400  10.96295
#> 2007-02-15  13.567  13.596  13.462  12.17286  90915300  10.95138
#> 2007-02-16  13.537  13.562  13.443  12.11857  99967000  10.90255
#> 2007-02-20  13.442  13.681  13.364  12.27143 154425600  11.04007
#> 2007-02-21  13.653  14.210  13.650  12.74286 288828400  11.46419
#> 2007-02-22  14.418  14.420  14.058  12.78714 209556200  11.50403
#> 2007-02-23  14.158  14.345  14.109  12.72429 129473400  11.44748
#> 2007-02-26  14.266  14.291  13.912  12.64429 153962200  11.37551
#> 2007-02-27  13.704  13.828  13.245  11.99000 286453300  10.78688
#> 2007-02-28  13.180  13.592  13.180  12.08714 229868800  10.87427
#> 2007-03-01  13.343  14.023  13.299  12.43714 353882200  11.18915
#> 2007-03-02  13.778  13.901  13.531  12.20143 215000100  10.97709
#> 2007-03-05  13.639  14.077  13.618  12.33143 209724900  11.09405
#> 2007-03-06  13.942  14.023  13.878  12.59857 180796700  11.33438
#> 2007-03-07  13.982  14.128  13.886  12.53143 156571100  11.27397
#> 2007-03-08  14.067  14.088  13.888  12.57143 127752800  11.30996
#> 2007-03-09  14.101  14.109  13.878  12.56714 112959000  11.30611
#> 2007-03-12  13.985  14.290  13.972  12.83857 182352100  11.55030
#> 2007-03-13  14.197  14.386  14.037  12.62857 216972700  11.36137
#> 2007-03-14  14.069  14.291  13.961  12.85714 199146500  11.56700
#> 2007-03-15  14.285  14.348  14.182  12.79571 139874700  11.51174
#> 2007-03-16  14.218  14.290  14.183  12.79857 142926000  11.51431
#> 2007-03-19  14.329  14.537  14.226  13.01857 178240300  11.71223
#> 2007-03-20  14.506  14.583  14.459  13.06857 122229100  11.75722
#> 2007-03-21  14.607  14.926  14.553  13.41000 171724000  12.06439
#> 2007-03-22  14.883  14.984  14.768  13.42286 140373100  12.07595
#> 2007-03-23  14.823  14.937  14.815  13.36000 112721000  12.01940
#> 2007-03-26  14.925  15.228  14.815  13.69286 216246800  12.31886
#> 2007-03-27  15.198  15.376  15.085  13.63714 233013200  12.26874
#> 2007-03-28  15.066  15.149  14.791  13.32000 235584300  11.98342
#> 2007-03-29  14.957  14.957  14.645  13.39286 181430900  12.04897
#> 2007-03-30  14.971  15.034  14.728  13.27286 150139500  11.94100
#> 2007-04-02  14.949  14.966  14.771  13.37857 125498100  12.03611
#> 2007-04-03  14.949  15.122  14.888  13.50000 145983600  12.14536
#> 2007-04-04  15.076  15.107  14.947  13.46714 119196000  12.11580
#> 2007-04-05  14.945  15.034  14.850  13.52571  88879000  12.16849
#> 2007-04-09  15.118  15.133  14.774  13.37857 103335400  12.03611
#> 2007-04-10  14.874  14.968  14.833  13.46429  88116700  12.11323
#> 2007-04-11  14.910  14.918  14.661  13.22714 137254600  11.89988
#> 2007-04-12  14.615  14.658  14.406  13.17000 164168900  11.84847
#> 2007-04-13  14.434  14.513  14.301  12.89143 179985400  11.59785
#> 2007-04-16  14.382  14.529  14.331  13.06143 152258400  11.75079
#> 2007-04-17  14.609  14.656  14.244  12.90714 187980100  11.61199
#> 2007-04-18  14.317  14.426  14.228  12.91429 116011000  11.61841
#> 2007-04-19  14.321  14.490  14.264  12.89571 106478400  11.60171
#> 2007-04-20  14.432  14.479  14.379  12.99571 130694900  11.69167
#> 2007-04-23  14.544  14.895  14.517  13.35857 195072500  12.01812
#> 2007-04-24  14.920  15.306  14.498  13.32000 263813200  11.98342
#> 2007-04-25  14.963  15.149  14.895  13.62143 296786000  12.25460
#> 2007-04-26  16.130  16.276  15.609  14.12000 434444500  12.70314
#> 2007-04-27  15.590  15.871  15.512  14.27429 174850900  12.84195
#> 2007-04-30  15.893  16.038  15.827  14.25714 154127400  12.82653
#> 2007-05-01  15.814  15.935  15.649  14.21000 133130900  12.78411
#> 2007-05-02  15.824  15.965  15.795  14.34143 126286300  12.90235
#> 2007-05-03  15.995  16.109  15.881  14.34286 144019400  12.90364
#> 2007-05-04  16.006  16.133  15.958  14.40143  95496800  12.95633
#> 2007-05-07  16.051  16.570  16.039  14.84572 215389300  13.35604
#> 2007-05-08  16.430  16.697  16.422  15.00857 195999300  13.50255
#> 2007-05-09  16.659  16.984  16.656  15.26857 179439400  13.73646
#> 2007-05-10  16.932  17.283  16.819  15.33429 299314400  13.79558
#> 2007-05-11  17.108  17.329  16.956  15.53429 163424100  13.97551
#> 2007-05-14  17.407  17.467  17.189  15.62286 162986600  14.05520
#> 2007-05-15  17.399  17.499  16.908  15.36000 238628600  13.81871
#> 2007-05-16  17.234  17.281  16.422  15.33429 281691900  13.79558
#> 2007-05-17  17.014  17.446  17.014  15.63429 183822800  14.06548
#> 2007-05-18  17.504  17.569  17.430  15.71714 155336300  14.14002
#> 2007-05-21  17.516  17.856  17.475  15.99714 159973100  14.39193
#> 2007-05-22  17.862  18.062  17.786  16.22000 143102400  14.59242
#> 2007-05-23  18.105  18.261  17.878  16.12714 227843700  14.50888
#> 2007-05-24  17.913  18.175  17.526  15.81286 221840500  14.22613
#> 2007-05-25  17.785  18.067  17.705  16.23143 158239900  14.60270
#> 2007-05-29  18.174  18.239  17.894  16.33571 161423500  14.69652
#> 2007-05-30  18.150  18.877  18.028  16.96714 369611200  15.26459
#> 2007-05-31  19.066  19.399  18.982  17.31286 324266600  15.57562
#> 2007-06-01  19.230  19.244  18.783  16.91429 221315500  15.21704
#> 2007-06-04  18.837  19.330  18.721  17.33286 221668300  15.59361
#> 2007-06-05  19.279  19.482  19.134  17.52429 230196400  15.76583
#> 2007-06-06  19.420  19.698  19.365  17.66286 278060300  15.89050
#> 2007-06-07  19.847  20.263  19.561  17.72428 478769900  15.94576
#> 2007-06-08  19.979  19.981  19.419  17.78429 310420600  15.99974
#> 2007-06-11  20.008  20.031  18.982  17.17000 468564600  15.44709
#> 2007-06-12  18.952  19.326  18.787  17.19714 356641600  15.47151
#> 2007-06-13  19.238  19.244  18.324  16.78571 430338300  15.10137
#> 2007-06-14  18.610  18.968  18.486  16.96429 243316500  15.26202
#> 2007-06-15  19.153  19.161  19.033  17.21429 202804700  15.48694
#> 2007-06-18  19.576  19.877  19.458  17.87000 227651200  16.07685
#> 2007-06-19  19.800  19.850  19.517  17.66571 235756500  15.89307
#> 2007-06-20  19.669  19.795  19.293  17.36429 224378000  15.62188
#> 2007-06-21  19.325  19.736  19.169  17.70000 216761300  15.92391
#> 2007-06-22  19.666  19.762  19.433  17.57143 157969000  15.80824
#> 2007-06-25  19.720  19.863  19.223  17.47714 241350900  15.72342
#> 2007-06-26  19.687  19.690  18.852  17.09286 336251300  15.37769
#> 2007-06-27  19.152  19.379  18.937  17.41286 243674200  15.66558
#> 2007-06-28  19.430  19.450  19.055  17.22286 209535900  15.49465
#> 2007-06-29  19.368  19.690  19.228  17.43429 284460400  15.68486
#> 2007-07-02  19.222  19.387  18.944  17.32286 248715600  15.58461
#> 2007-07-03  19.372  20.230  19.293  18.16714 290620400  16.34418
#> 2007-07-05  20.452  21.114  20.435  18.96429 363262900  17.06133
#> 2007-07-06  21.140  21.173  20.706  18.90000 218673700  17.00350
#> 2007-07-09  21.021  21.103  20.513  18.61857 248955000  16.75031
#> 2007-07-10  20.465  21.357  20.454  18.90714 313751900  17.00993
#> 2007-07-11  20.972  21.230  20.851  18.91286 205443000  17.01507
#> 2007-07-12  21.254  21.316  21.022  19.15286 176152200  17.23098
#> 2007-07-13  21.442  21.889  21.361  19.67571 226901500  17.70138
#> 2007-07-16  21.975  22.228  21.834  19.72857 234028200  17.74893
#> 2007-07-17  21.961  22.167  21.834  19.84429 177489900  17.85303
#> 2007-07-18  21.943  21.983  21.602  19.73143 189214200  17.75150
#> 2007-07-19  22.278  22.359  22.175  20.00000 183222900  17.99312
#> 2007-07-20  22.493  22.894  22.231  20.53571 291943400  18.47508
#> 2007-07-23  22.756  23.060  22.378  20.52857 259122500  18.46865
#> 2007-07-24  22.053  22.390  21.302  19.27000 448823200  17.33637
#> 2007-07-25  21.810  21.970  21.437  19.60857 374045700  17.64097
#> 2007-07-26  23.169  23.580  21.748  20.85714 546657300  18.76425
#> 2007-07-27  23.214  23.647  22.831  20.55000 290274600  18.48793
#> 2007-07-30  22.918  23.096  22.162  20.20429 276747100  18.17691
#> 2007-07-31  22.702  22.783  20.884  18.82286 440598200  16.93409
#> 2007-08-01  21.221  21.497  20.289  19.28571 437539200  17.35051
#> 2007-08-02  21.699  21.748  21.302  19.49857 213161200  17.54201
#> 2007-08-03  21.478  21.588  20.881  18.83571 169796900  16.94567
#> 2007-08-06  21.103  21.480  20.373  19.32143 231292600  17.38263
#> 2007-08-07  21.427  21.792  21.060  19.29000 237484100  17.35437
#> 2007-08-08  21.716  21.732  20.960  19.14428 202024200  17.22328
#> 2007-08-09  20.819  21.119  19.863  18.05571 281348900  16.24393
#> 2007-08-10  19.550  20.286  19.103  17.85714 352687300  16.06528
#> 2007-08-13  20.376  20.540  20.087  18.25571 188227900  16.42387
#> 2007-08-14  20.371  20.373  19.644  17.71857 184751700  15.94062
#> 2007-08-15  19.490  19.827  18.999  17.12857 248213000  15.40982
#> 2007-08-16  18.580  18.817  17.724  16.72143 466672500  15.04353
#> 2007-08-17  19.374  19.611  19.026  17.43714 298765600  15.68743
#> 2007-08-20  19.684  19.769  19.134  17.46000 200829300  15.70799
#> 2007-08-21  19.406  20.478  19.214  18.22428 325761800  16.39559
#> 2007-08-22  20.837  21.079  20.695  18.93000 265441400  17.03049
#> 2007-08-23  21.133  21.173  20.605  18.72428 216709500  16.84541
#> 2007-08-24  20.727  21.496  20.613  19.32857 227958500  17.38906
#> 2007-08-27  21.181  21.383  20.976  18.89286 176859900  16.99707
#> 2007-08-28  20.800  21.025  20.108  18.11714 294841400  16.29920
#> 2007-08-29  20.624  21.307  20.570  19.15429 291715200  17.23226
#>  [ reached getOption("max.print") -- omitted 2509 rows ]