Summarise each group by time

summarise_by_time() Is a time-series variant of the popular dplyr::summarise() function.

summarise_by_time() and summarize_by_time() are synonyms.

summarise_by_time(
  .data,
  .date_var,
  ...,
  .by = "week",
  .type = c("floor", "ceiling", "round")
)

summarize_by_time(
  .data,
  .date_var,
  ...,
  .by = "week",
  .type = c("floor", "ceiling", "round")
)

Arguments

.data
A tbl object or data.frame

.date_var
A column of date or date-time (e.g. POSIXct) data class

...
Name-value pairs of summary functions. The name will be the name of the variable in the result.

The value can be:

A vector of length 1, e.g. min(x), n(), or sum(is.na(y)).

A vector of length n, e.g. quantile().

A data frame, to add multiple columns from a single expression.

.by
A time unit to summarise by. Time units are collapsed using lubridate::floor_date() or lubridate::ceiling_date().

The value can be:

second

minute

hour

day

week

month

bimonth

quarter

season

halfyear

year

Arbitrary unique English abbreviations as in the lubridate::period() constructor are allowed.

.type
One of "floor", "ceiling", or "round. Defaults to "floor". See lubridate::round_date.

Value

An object usually of the same type as .data.

The rows come from the underlying group_keys().

The columns are a combination of the grouping keys and the summary expressions that you provide.

If x is grouped by more than one variable, the output will be another grouped_df with the right-most group removed.

If x is grouped by one variable, or is not grouped, the output will be a tibble.

Data frame attributes are not preserved, because summarise() fundamentally creates a new data frame.

Useful summary functions

Sum: SUM()

Center: AVERAGE(), MEDIAN()

Spread: STDEV(), VAR()

Range: MIN(), MAX()

Count: COUNT(), COUNT_UNIQUE()

Position: FIRST(), LAST(), NTH()

Correlation: COR(), COV()

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

Examples

# Libraries
library(tidyquant)
library(dplyr)

# First adjusted price in each month
FANG %>%
    group_by(symbol) %>%
    summarise_by_time(
        .date_var  = date,
        .by         = "month",
        adjusted   = FIRST(adjusted)
    )

#> # A tibble: 192 x 3
#> # Groups:   symbol [4]
#>    symbol date       adjusted
#>    <chr>  <date>        <dbl>
#>  1 AMZN   2013-01-01     257.
#>  2 AMZN   2013-02-01     265 
#>  3 AMZN   2013-03-01     266.
#>  4 AMZN   2013-04-01     262.
#>  5 AMZN   2013-05-01     248.
#>  6 AMZN   2013-06-01     267.
#>  7 AMZN   2013-07-01     282.
#>  8 AMZN   2013-08-01     306.
#>  9 AMZN   2013-09-01     289.
#> 10 AMZN   2013-10-01     321.
#> # … with 182 more rows

# Last adjused price in each month (day is last day of month with ceiling option)
FANG %>%
    group_by(symbol) %>%
    summarise_by_time(
        .date_var  = date,
        .by        = "month",
        adjusted   = LAST(adjusted),
        .type      = "ceiling")

#> # A tibble: 192 x 3
#> # Groups:   symbol [4]
#>    symbol date       adjusted
#>    <chr>  <date>        <dbl>
#>  1 AMZN   2013-02-01     266.
#>  2 AMZN   2013-03-01     264.
#>  3 AMZN   2013-04-01     266.
#>  4 AMZN   2013-05-01     254.
#>  5 AMZN   2013-06-01     269.
#>  6 AMZN   2013-07-01     278.
#>  7 AMZN   2013-08-01     301.
#>  8 AMZN   2013-09-01     281.
#>  9 AMZN   2013-10-01     313.
#> 10 AMZN   2013-11-01     364.
#> # … with 182 more rows

# Total Volume each year (.by is set to "year" now)
FANG %>%
    group_by(symbol) %>%
    summarise_by_time(
        .date_var  = date,
        .by        = "year",
        adjusted   = SUM(volume))

#> # A tibble: 16 x 3
#> # Groups:   symbol [4]
#>    symbol date          adjusted
#>    <chr>  <date>           <dbl>
#>  1 AMZN   2013-01-01   747905700
#>  2 AMZN   2014-01-01  1029066700
#>  3 AMZN   2015-01-01   956936800
#>  4 AMZN   2016-01-01  1037105600
#>  5 FB     2013-01-01 15143182600
#>  6 FB     2014-01-01 11977699100
#>  7 FB     2015-01-01  6792708200
#>  8 FB     2016-01-01  6414357000
#>  9 GOOG   2013-01-01  1055967100
#> 10 GOOG   2014-01-01   626733500
#> 11 GOOG   2015-01-01   521446300
#> 12 GOOG   2016-01-01   461120900
#> 13 NFLX   2013-01-01  6915790700
#> 14 NFLX   2014-01-01  4898415200
#> 15 NFLX   2015-01-01  4679881700
#> 16 NFLX   2016-01-01  3228349500