Coerce to tibble. Enable preserving row names when coercing matrix and time-series-like objects with row names.

Coerce to tibble. Enable preserving row names when coercing matrix and time-series-like objects with row names.

as_tibble(x, preserve_row_names = FALSE, ...)

Arguments

x
A list, matrix, xts, zoo, timeSeries, etc object.

preserve_row_names
Used during coercion from matrix, xts, zoo, timeSeries, etc objects that have row names. When TRUE, creates a row.names column with names of rows as character class.

...
Additional parameters passed to the appropriate tibble::as_tibble() function.

Value

Returns a tibble object.

Details

as_tibble is a wrapper for tibble::as_tibble that includes a preserve_row_names argument. The function is designed to coerce xts, zoo, timeSeries, ts, and irts objects that are used frequently in quantitative financial analysis. When preserve_row_names = TRUE is specified, a new column, row.names, is created during object coercion as a character class.

See also

as_xts()

Examples

# Load libraries
library(tidyverse)
library(tidyquant)

# Matrix coercion to tibble
m <- matrix(rnorm(50), ncol = 5)
colnames(m) <- c("a", "b", "c", "d", "e")
rownames(m) <- letters[1:nrow(m)]
m_tbl <- as_tibble(m, preserve_row_names = TRUE)

# xts coercion to tibble
quantmod::getSymbols("AAPL", auto.assign = FALSE) %>%
    as_tibble(preserve_row_names = TRUE)

#>     As of 0.4-0, ‘getSymbols’ uses env=parent.frame() and
#>  auto.assign=TRUE by default.
#> 
#>  This  behavior  will be  phased out in 0.5-0  when the call  will
#>  default to use auto.assign=FALSE. getOption("getSymbols.env") and 
#>  getOptions("getSymbols.auto.assign") are now checked for alternate defaults
#> 
#>  This message is shown once per session and may be disabled by setting 
#>  options("getSymbols.warning4.0"=FALSE). See ?getSymbols for more details.

#> Warning: the condition has length > 1 and only the first element will be used

#> # A tibble: 2,591 × 7
#>     row.names AAPL.Open AAPL.High AAPL.Low AAPL.Close AAPL.Volume AAPL.Adjusted
#>         <chr>     <dbl>     <dbl>    <dbl>      <dbl>       <dbl>         <dbl>
#>  1 2007-01-03     86.29     86.58    81.90      83.80   309579900      10.85709
#>  2 2007-01-04     84.05     85.95    83.82      85.66   211815100      11.09807
#>  3 2007-01-05     85.77     86.20    84.40      85.05   208685400      11.01904
#>  4 2007-01-08     85.96     86.53    85.28      85.47   199276700      11.07345
#>  5 2007-01-09     86.45     92.98    85.15      92.57   837324600      11.99333
#>  6 2007-01-10     94.75     97.80    93.45      97.00   738220000      12.56728
#>  7 2007-01-11     95.94     96.78    95.10      95.80   360063200      12.41180
#>  8 2007-01-12     94.59     95.06    93.23      94.62   328172600      12.25892
#>  9 2007-01-16     95.68     97.25    95.45      97.10   311019100      12.58023
#> 10 2007-01-17     97.56     97.60    94.82      94.95   411565000      12.30168
#> # ... with 2,581 more rows