Coerce objects to xts, designed to work with tibble and data.frame objects

Coerce objects to xts, designed to work with tibble and data.frame objects

as_xts(x, date_col = NULL, ...)

as_xts_(x, date_col = NULL, ...)

Arguments

x
A data.frame (with date column), matrix, xts, zoo, timeSeries, etc object.

date_col
Required for objects that inherit the data.frame class. Must specify a date column that is of the date class. Unused for non-data.frame objects.

...
Additional parameters passed to xts::as.xts.

Value

Returns a xts object.

Details

as_xts is a wrapper for xts::as.xts that includes a date_col argument. When date_col is specified, the date column is used as row names during coercion to xts class. The date column must be in a date format (i.e. date class). as_xts_ evaluates the date_col using Non-Standard Evaluation (NSE). See vignette("nse") for details.

It is possible to coerce non-data.frame-like objects including zoo, timeSeries, ts, and irts objects. There is no need to specify the date_col argument.

See also

as_tibble()

Examples

# Load libraries
library(tidyverse)
library(tidyquant)

tq_get("AAPL", get = "stock.prices") %>%
    as_xts(date_col = date)

#>              open   high    low  close    volume  adjusted
#> 2007-01-03  86.29  86.58  81.90  83.80 309579900  10.85709
#> 2007-01-04  84.05  85.95  83.82  85.66 211815100  11.09807
#> 2007-01-05  85.77  86.20  84.40  85.05 208685400  11.01904
#> 2007-01-08  85.96  86.53  85.28  85.47 199276700  11.07345
#> 2007-01-09  86.45  92.98  85.15  92.57 837324600  11.99333
#> 2007-01-10  94.75  97.80  93.45  97.00 738220000  12.56728
#> 2007-01-11  95.94  96.78  95.10  95.80 360063200  12.41180
#> 2007-01-12  94.59  95.06  93.23  94.62 328172600  12.25892
#> 2007-01-16  95.68  97.25  95.45  97.10 311019100  12.58023
#> 2007-01-17  97.56  97.60  94.82  94.95 411565000  12.30168
#> 2007-01-18  92.10  92.11  89.05  89.07 591151400  11.53987
#> 2007-01-19  88.63  89.65  88.12  88.50 341118400  11.46602
#> 2007-01-22  89.14  89.16  85.65  86.79 363506500  11.24447
#> 2007-01-23  85.73  87.51  85.51  85.70 301856100  11.10325
#> 2007-01-24  86.68  87.15  86.08  86.70 231953400  11.23281
#> 2007-01-25  87.11  88.50  86.03  86.25 226493400  11.17451
#> 2007-01-26  87.11  87.37  84.99  85.38 246718500  11.06179
#> 2007-01-29  86.30  86.65  85.53  85.94 225416100  11.13435
#> 2007-01-30  86.43  86.49  85.25  85.55 144492600  11.08382
#> 2007-01-31  84.86  86.00  84.35  85.73 214017300  11.10714
#> 2007-02-01  86.23  86.27  84.74  84.74 166085500  10.97888
#> 2007-02-02  84.12  85.25  83.70  84.75 155382500  10.98017
#> 2007-02-05  84.30  85.23  83.94  83.94 144713100  10.87523
#> 2007-02-06  84.45  84.47  82.86  84.15 216098400  10.90244
#> 2007-02-07  84.48  86.38  83.55  86.15 266706300  11.16156
#> 2007-02-08  85.43  86.51  85.41  86.18 169757700  11.16544
#> 2007-02-09  85.88  86.20  83.21  83.27 215135200  10.78842
#> 2007-02-12  84.43  85.18  83.63  84.88 181017900  10.99701
#> 2007-02-13  85.16  85.29  84.30  84.70 145246500  10.97369
#> 2007-02-14  84.63  85.64  84.57  85.30 126995400  11.05143
#> 2007-02-15  85.44  85.62  84.78  85.21  90915300  11.03977
#> 2007-02-16  85.25  85.41  84.66  84.83  99967000  10.99054
#> 2007-02-20  84.65  86.16  84.16  85.90 154425600  11.12917
#> 2007-02-21  85.98  89.49  85.96  89.20 288828400  11.55671
#> 2007-02-22  90.80  90.81  88.53  89.51 209556200  11.59688
#> 2007-02-23  89.16  90.34  88.85  89.07 129473400  11.53987
#> 2007-02-26  89.84  90.00  87.61  88.51 153962200  11.46732
#> 2007-02-27  86.30  87.08  83.41  83.93 286453300  10.87393
#> 2007-02-28  83.00  85.60  83.00  84.61 229868800  10.96203
#> 2007-03-01  84.03  88.31  83.75  87.06 353882200  11.27946
#> 2007-03-02  86.77  87.54  85.21  85.41 215000100  11.06568
#> 2007-03-05  85.89  88.65  85.76  86.32 209724900  11.18358
#> 2007-03-06  87.80  88.31  87.40  88.19 180796700  11.42586
#> 2007-03-07  88.05  88.97  87.45  87.72 156571100  11.36496
#> 2007-03-08  88.59  88.72  87.46  88.00 127752800  11.40124
#> 2007-03-09  88.80  88.85  87.40  87.97 112959000  11.39735
#> 2007-03-12  88.07  89.99  87.99  89.87 182352100  11.64352
#> 2007-03-13  89.41  90.60  88.40  88.40 216972700  11.45306
#> 2007-03-14  88.60  90.00  87.92  90.00 199146500  11.66036
#> 2007-03-15  89.96  90.36  89.31  89.57 139874700  11.60465
#> 2007-03-16  89.54  89.99  89.32  89.59 142926000  11.60724
#> 2007-03-19  90.24  91.55  89.59  91.13 178240300  11.80676
#> 2007-03-20  91.35  91.84  91.06  91.48 122229100  11.85211
#> 2007-03-21  91.99  94.00  91.65  93.87 171724000  12.16175
#> 2007-03-22  93.73  94.36  93.00  93.96 140373100  12.17342
#> 2007-03-23  93.35  94.07  93.30  93.52 112721000  12.11641
#> 2007-03-26  93.99  95.90  93.30  95.85 216246800  12.41828
#> 2007-03-27  95.71  96.83  95.00  95.46 233013200  12.36776
#> 2007-03-28  94.88  95.40  93.15  93.24 235584300  12.08013
#> 2007-03-29  94.19  94.19  92.23  93.75 181430900  12.14621
#> 2007-03-30  94.28  94.68  92.75  92.91 150139500  12.03738
#> 2007-04-02  94.14  94.25  93.02  93.65 125498100  12.13325
#> 2007-04-03  94.14  95.23  93.76  94.50 145983600  12.24338
#> 2007-04-04  94.94  95.14  94.13  94.27 119196000  12.21358
#> 2007-04-05  94.12  94.68  93.52  94.68  88879000  12.26670
#> 2007-04-09  95.21  95.30  93.04  93.65 103335400  12.13325
#> 2007-04-10  93.67  94.26  93.41  94.25  88116700  12.21099
#> 2007-04-11  93.90  93.95  92.33  92.59 137254600  11.99592
#> 2007-04-12  92.04  92.31  90.72  92.19 164168900  11.94410
#> 2007-04-13  90.90  91.40  90.06  90.24 179985400  11.69145
#> 2007-04-16  90.57  91.50  90.25  91.43 152258400  11.84563
#> 2007-04-17  92.00  92.30  89.70  90.35 187980100  11.70571
#> 2007-04-18  90.16  90.85  89.60  90.40 116011000  11.71218
#> 2007-04-19  90.19  91.25  89.83  90.27 106478400  11.69534
#> 2007-04-20  90.89  91.18  90.55  90.97 130694900  11.78603
#> 2007-04-23  91.59  93.80  91.42  93.51 195072500  12.11511
#> 2007-04-24  93.96  96.39  91.30  93.24 263813200  12.08013
#> 2007-04-25  94.23  95.40  93.80  95.35 296786000  12.35350
#> 2007-04-26 101.58 102.50  98.30  98.84 434444500  12.80567
#> 2007-04-27  98.18  99.95  97.69  99.92 174850900  12.94559
#> 2007-04-30 100.09 101.00  99.67  99.80 154127400  12.93004
#> 2007-05-01  99.59 100.35  98.55  99.47 133130900  12.88729
#> 2007-05-02  99.65 100.54  99.47 100.39 126286300  13.00648
#> 2007-05-03 100.73 101.45 100.01 100.40 144019400  13.00778
#> 2007-05-04 100.80 101.60 100.50 100.81  95496800  13.06090
#> 2007-05-07 101.08 104.35 101.01 103.92 215389300  13.46383
#> 2007-05-08 103.47 105.15 103.42 105.06 195999300  13.61153
#> 2007-05-09 104.91 106.96 104.89 106.88 179439400  13.84732
#> 2007-05-10 106.63 108.84 105.92 107.34 299314400  13.90692
#> 2007-05-11 107.74 109.13 106.78 108.74 163424100  14.08831
#> 2007-05-14 109.62 110.00 108.25 109.36 162986600  14.16863
#> 2007-05-15 109.57 110.20 106.48 107.52 238628600  13.93024
#> 2007-05-16 108.53 108.83 103.42 107.34 281691900  13.90692
#> 2007-05-17 107.15 109.87 107.15 109.44 183822800  14.17900
#> 2007-05-18 110.23 110.64 109.77 110.02 155336300  14.25414
#> 2007-05-21 110.31 112.45 110.05 111.98 159973100  14.50808
#> 2007-05-22 112.49 113.75 112.01 113.54 143102400  14.71019
#> 2007-05-23 114.02 115.00 112.59 112.89 227843700  14.62598
#> 2007-05-24 112.81 114.46 110.37 110.69 221840500  14.34095
#> 2007-05-25 112.00 113.78 111.50 113.62 158239900  14.72056
#> 2007-05-29 114.45 114.86 112.69 114.35 161423500  14.81513
#> 2007-05-30 114.30 118.88 113.53 118.77 369611200  15.38779
#> 2007-05-31 120.07 122.17 119.54 121.19 324266600  15.70132
#> 2007-06-01 121.10 121.19 118.29 118.40 221315500  15.33985
#> 2007-06-04 118.63 121.73 117.90 121.33 221668300  15.71946
#> 2007-06-05 121.41 122.69 120.50 122.67 230196400  15.89307
#> 2007-06-06 122.30 124.05 121.95 123.64 278060300  16.01874
#> 2007-06-07 124.99 127.61 123.19 124.07 478769900  16.07445
#> 2007-06-08 125.82 125.83 122.29 124.49 310420600  16.12887
#> 2007-06-11 126.00 126.15 119.54 120.19 468564600  15.57176
#> 2007-06-12 119.35 121.71 118.31 120.38 356641600  15.59638
#> 2007-06-13 121.15 121.19 115.40 117.50 430338300  15.22325
#> 2007-06-14 117.20 119.45 116.42 118.75 243316500  15.38520
#> 2007-06-15 120.62 120.67 119.86 120.50 202804700  15.61192
#> 2007-06-18 123.28 125.18 122.54 125.09 227651200  16.20660
#> 2007-06-19 124.69 125.01 122.91 123.66 235756500  16.02133
#> 2007-06-20 123.87 124.66 121.50 121.55 224378000  15.74796
#> 2007-06-21 121.70 124.29 120.72 123.90 216761300  16.05243
#> 2007-06-22 123.85 124.45 122.38 123.00 157969000  15.93582
#> 2007-06-25 124.19 125.09 121.06 122.34 241350900  15.85032
#> 2007-06-26 123.98 124.00 118.72 119.65 336251300  15.50180
#> 2007-06-27 120.61 122.04 119.26 121.89 243674200  15.79201
#> 2007-06-28 122.36 122.49 120.00 120.56 209535900  15.61970
#> 2007-06-29 121.97 124.00 121.09 122.04 284460400  15.81145
#> 2007-07-02 121.05 122.09 119.30 121.26 248715600  15.71039
#> 2007-07-03 122.00 127.40 121.50 127.17 290620400  16.47609
#> 2007-07-05 128.80 132.97 128.69 132.75 363262900  17.19903
#> 2007-07-06 133.13 133.34 130.40 132.30 218673700  17.14073
#> 2007-07-09 132.38 132.90 129.18 130.33 248955000  16.88550
#> 2007-07-10 128.88 134.50 128.81 132.35 313751900  17.14721
#> 2007-07-11 132.07 133.70 131.31 132.39 205443000  17.15239
#> 2007-07-12 133.85 134.24 132.39 134.07 176152200  17.37005
#> 2007-07-13 135.03 137.85 134.52 137.73 226901500  17.84424
#> 2007-07-16 138.39 139.98 137.50 138.10 234028200  17.89217
#> 2007-07-17 138.30 139.60 137.50 138.91 177489900  17.99712
#> 2007-07-18 138.19 138.44 136.04 138.12 189214200  17.89476
#> 2007-07-19 140.30 140.81 139.65 140.00 183222900  18.13834
#> 2007-07-20 141.65 144.18 140.00 143.75 291943400  18.62419
#> 2007-07-23 143.31 145.22 140.93 143.70 259122500  18.61771
#> 2007-07-24 138.88 141.00 134.15 134.89 448823200  17.47629
#> 2007-07-25 137.35 138.36 135.00 137.26 374045700  17.78334
#> 2007-07-26 145.91 148.50 136.96 146.00 546657300  18.91569
#> 2007-07-27 146.19 148.92 143.78 143.85 290274600  18.63714
#> 2007-07-30 144.33 145.45 139.57 141.43 276747100  18.32361
#> 2007-07-31 142.97 143.48 131.52 131.76 440598200  17.07077
#> 2007-08-01 133.64 135.38 127.77 135.00 437539200  17.49054
#> 2007-08-02 136.65 136.96 134.15 136.49 213161200  17.68358
#> 2007-08-03 135.26 135.95 131.50 131.85 169796900  17.08243
#> 2007-08-06 132.90 135.27 128.30 135.25 231292600  17.52293
#> 2007-08-07 134.94 137.24 132.63 135.03 237484100  17.49443
#> 2007-08-08 136.76 136.86 132.00 134.01 202024200  17.36228
#> 2007-08-09 131.11 133.00 125.09 126.39 281348900  16.37503
#> 2007-08-10 123.12 127.75 120.30 125.00 352687300  16.19494
#> 2007-08-13 128.32 129.35 126.50 127.79 188227900  16.55642
#> 2007-08-14 128.29 128.30 123.71 124.03 184751700  16.06927
#> 2007-08-15 122.74 124.86 119.65 119.90 248213000  15.53419
#> 2007-08-16 117.01 118.50 111.62 117.05 466672500  15.16494
#> 2007-08-17 122.01 123.50 119.82 122.06 298765600  15.81404
#> 2007-08-20 123.96 124.50 120.50 122.22 200829300  15.83477
#> 2007-08-21 122.21 128.96 121.00 127.57 325761800  16.52791
#> 2007-08-22 131.22 132.75 130.33 132.51 265441400  17.16794
#> 2007-08-23 133.09 133.34 129.76 131.07 216709500  16.98137
#> 2007-08-24 130.53 135.37 129.81 135.30 227958500  17.52941
#> 2007-08-27 133.39 134.66 132.10 132.25 176859900  17.13425
#> 2007-08-28 130.99 132.41 126.63 126.82 294841400  16.43074
#> 2007-08-29 129.88 134.18 129.54 134.08 291715200  17.37134
#>  [ reached getOption("max.print") -- omitted 2419 rows ]

# Non-Standard Evaluation (NSE)
x <- "date"
tq_get("AAPL", get = "stock.prices") %>%
    as_xts_(date_col = x)

#>              open   high    low  close    volume  adjusted
#> 2007-01-03  86.29  86.58  81.90  83.80 309579900  10.85709
#> 2007-01-04  84.05  85.95  83.82  85.66 211815100  11.09807
#> 2007-01-05  85.77  86.20  84.40  85.05 208685400  11.01904
#> 2007-01-08  85.96  86.53  85.28  85.47 199276700  11.07345
#> 2007-01-09  86.45  92.98  85.15  92.57 837324600  11.99333
#> 2007-01-10  94.75  97.80  93.45  97.00 738220000  12.56728
#> 2007-01-11  95.94  96.78  95.10  95.80 360063200  12.41180
#> 2007-01-12  94.59  95.06  93.23  94.62 328172600  12.25892
#> 2007-01-16  95.68  97.25  95.45  97.10 311019100  12.58023
#> 2007-01-17  97.56  97.60  94.82  94.95 411565000  12.30168
#> 2007-01-18  92.10  92.11  89.05  89.07 591151400  11.53987
#> 2007-01-19  88.63  89.65  88.12  88.50 341118400  11.46602
#> 2007-01-22  89.14  89.16  85.65  86.79 363506500  11.24447
#> 2007-01-23  85.73  87.51  85.51  85.70 301856100  11.10325
#> 2007-01-24  86.68  87.15  86.08  86.70 231953400  11.23281
#> 2007-01-25  87.11  88.50  86.03  86.25 226493400  11.17451
#> 2007-01-26  87.11  87.37  84.99  85.38 246718500  11.06179
#> 2007-01-29  86.30  86.65  85.53  85.94 225416100  11.13435
#> 2007-01-30  86.43  86.49  85.25  85.55 144492600  11.08382
#> 2007-01-31  84.86  86.00  84.35  85.73 214017300  11.10714
#> 2007-02-01  86.23  86.27  84.74  84.74 166085500  10.97888
#> 2007-02-02  84.12  85.25  83.70  84.75 155382500  10.98017
#> 2007-02-05  84.30  85.23  83.94  83.94 144713100  10.87523
#> 2007-02-06  84.45  84.47  82.86  84.15 216098400  10.90244
#> 2007-02-07  84.48  86.38  83.55  86.15 266706300  11.16156
#> 2007-02-08  85.43  86.51  85.41  86.18 169757700  11.16544
#> 2007-02-09  85.88  86.20  83.21  83.27 215135200  10.78842
#> 2007-02-12  84.43  85.18  83.63  84.88 181017900  10.99701
#> 2007-02-13  85.16  85.29  84.30  84.70 145246500  10.97369
#> 2007-02-14  84.63  85.64  84.57  85.30 126995400  11.05143
#> 2007-02-15  85.44  85.62  84.78  85.21  90915300  11.03977
#> 2007-02-16  85.25  85.41  84.66  84.83  99967000  10.99054
#> 2007-02-20  84.65  86.16  84.16  85.90 154425600  11.12917
#> 2007-02-21  85.98  89.49  85.96  89.20 288828400  11.55671
#> 2007-02-22  90.80  90.81  88.53  89.51 209556200  11.59688
#> 2007-02-23  89.16  90.34  88.85  89.07 129473400  11.53987
#> 2007-02-26  89.84  90.00  87.61  88.51 153962200  11.46732
#> 2007-02-27  86.30  87.08  83.41  83.93 286453300  10.87393
#> 2007-02-28  83.00  85.60  83.00  84.61 229868800  10.96203
#> 2007-03-01  84.03  88.31  83.75  87.06 353882200  11.27946
#> 2007-03-02  86.77  87.54  85.21  85.41 215000100  11.06568
#> 2007-03-05  85.89  88.65  85.76  86.32 209724900  11.18358
#> 2007-03-06  87.80  88.31  87.40  88.19 180796700  11.42586
#> 2007-03-07  88.05  88.97  87.45  87.72 156571100  11.36496
#> 2007-03-08  88.59  88.72  87.46  88.00 127752800  11.40124
#> 2007-03-09  88.80  88.85  87.40  87.97 112959000  11.39735
#> 2007-03-12  88.07  89.99  87.99  89.87 182352100  11.64352
#> 2007-03-13  89.41  90.60  88.40  88.40 216972700  11.45306
#> 2007-03-14  88.60  90.00  87.92  90.00 199146500  11.66036
#> 2007-03-15  89.96  90.36  89.31  89.57 139874700  11.60465
#> 2007-03-16  89.54  89.99  89.32  89.59 142926000  11.60724
#> 2007-03-19  90.24  91.55  89.59  91.13 178240300  11.80676
#> 2007-03-20  91.35  91.84  91.06  91.48 122229100  11.85211
#> 2007-03-21  91.99  94.00  91.65  93.87 171724000  12.16175
#> 2007-03-22  93.73  94.36  93.00  93.96 140373100  12.17342
#> 2007-03-23  93.35  94.07  93.30  93.52 112721000  12.11641
#> 2007-03-26  93.99  95.90  93.30  95.85 216246800  12.41828
#> 2007-03-27  95.71  96.83  95.00  95.46 233013200  12.36776
#> 2007-03-28  94.88  95.40  93.15  93.24 235584300  12.08013
#> 2007-03-29  94.19  94.19  92.23  93.75 181430900  12.14621
#> 2007-03-30  94.28  94.68  92.75  92.91 150139500  12.03738
#> 2007-04-02  94.14  94.25  93.02  93.65 125498100  12.13325
#> 2007-04-03  94.14  95.23  93.76  94.50 145983600  12.24338
#> 2007-04-04  94.94  95.14  94.13  94.27 119196000  12.21358
#> 2007-04-05  94.12  94.68  93.52  94.68  88879000  12.26670
#> 2007-04-09  95.21  95.30  93.04  93.65 103335400  12.13325
#> 2007-04-10  93.67  94.26  93.41  94.25  88116700  12.21099
#> 2007-04-11  93.90  93.95  92.33  92.59 137254600  11.99592
#> 2007-04-12  92.04  92.31  90.72  92.19 164168900  11.94410
#> 2007-04-13  90.90  91.40  90.06  90.24 179985400  11.69145
#> 2007-04-16  90.57  91.50  90.25  91.43 152258400  11.84563
#> 2007-04-17  92.00  92.30  89.70  90.35 187980100  11.70571
#> 2007-04-18  90.16  90.85  89.60  90.40 116011000  11.71218
#> 2007-04-19  90.19  91.25  89.83  90.27 106478400  11.69534
#> 2007-04-20  90.89  91.18  90.55  90.97 130694900  11.78603
#> 2007-04-23  91.59  93.80  91.42  93.51 195072500  12.11511
#> 2007-04-24  93.96  96.39  91.30  93.24 263813200  12.08013
#> 2007-04-25  94.23  95.40  93.80  95.35 296786000  12.35350
#> 2007-04-26 101.58 102.50  98.30  98.84 434444500  12.80567
#> 2007-04-27  98.18  99.95  97.69  99.92 174850900  12.94559
#> 2007-04-30 100.09 101.00  99.67  99.80 154127400  12.93004
#> 2007-05-01  99.59 100.35  98.55  99.47 133130900  12.88729
#> 2007-05-02  99.65 100.54  99.47 100.39 126286300  13.00648
#> 2007-05-03 100.73 101.45 100.01 100.40 144019400  13.00778
#> 2007-05-04 100.80 101.60 100.50 100.81  95496800  13.06090
#> 2007-05-07 101.08 104.35 101.01 103.92 215389300  13.46383
#> 2007-05-08 103.47 105.15 103.42 105.06 195999300  13.61153
#> 2007-05-09 104.91 106.96 104.89 106.88 179439400  13.84732
#> 2007-05-10 106.63 108.84 105.92 107.34 299314400  13.90692
#> 2007-05-11 107.74 109.13 106.78 108.74 163424100  14.08831
#> 2007-05-14 109.62 110.00 108.25 109.36 162986600  14.16863
#> 2007-05-15 109.57 110.20 106.48 107.52 238628600  13.93024
#> 2007-05-16 108.53 108.83 103.42 107.34 281691900  13.90692
#> 2007-05-17 107.15 109.87 107.15 109.44 183822800  14.17900
#> 2007-05-18 110.23 110.64 109.77 110.02 155336300  14.25414
#> 2007-05-21 110.31 112.45 110.05 111.98 159973100  14.50808
#> 2007-05-22 112.49 113.75 112.01 113.54 143102400  14.71019
#> 2007-05-23 114.02 115.00 112.59 112.89 227843700  14.62598
#> 2007-05-24 112.81 114.46 110.37 110.69 221840500  14.34095
#> 2007-05-25 112.00 113.78 111.50 113.62 158239900  14.72056
#> 2007-05-29 114.45 114.86 112.69 114.35 161423500  14.81513
#> 2007-05-30 114.30 118.88 113.53 118.77 369611200  15.38779
#> 2007-05-31 120.07 122.17 119.54 121.19 324266600  15.70132
#> 2007-06-01 121.10 121.19 118.29 118.40 221315500  15.33985
#> 2007-06-04 118.63 121.73 117.90 121.33 221668300  15.71946
#> 2007-06-05 121.41 122.69 120.50 122.67 230196400  15.89307
#> 2007-06-06 122.30 124.05 121.95 123.64 278060300  16.01874
#> 2007-06-07 124.99 127.61 123.19 124.07 478769900  16.07445
#> 2007-06-08 125.82 125.83 122.29 124.49 310420600  16.12887
#> 2007-06-11 126.00 126.15 119.54 120.19 468564600  15.57176
#> 2007-06-12 119.35 121.71 118.31 120.38 356641600  15.59638
#> 2007-06-13 121.15 121.19 115.40 117.50 430338300  15.22325
#> 2007-06-14 117.20 119.45 116.42 118.75 243316500  15.38520
#> 2007-06-15 120.62 120.67 119.86 120.50 202804700  15.61192
#> 2007-06-18 123.28 125.18 122.54 125.09 227651200  16.20660
#> 2007-06-19 124.69 125.01 122.91 123.66 235756500  16.02133
#> 2007-06-20 123.87 124.66 121.50 121.55 224378000  15.74796
#> 2007-06-21 121.70 124.29 120.72 123.90 216761300  16.05243
#> 2007-06-22 123.85 124.45 122.38 123.00 157969000  15.93582
#> 2007-06-25 124.19 125.09 121.06 122.34 241350900  15.85032
#> 2007-06-26 123.98 124.00 118.72 119.65 336251300  15.50180
#> 2007-06-27 120.61 122.04 119.26 121.89 243674200  15.79201
#> 2007-06-28 122.36 122.49 120.00 120.56 209535900  15.61970
#> 2007-06-29 121.97 124.00 121.09 122.04 284460400  15.81145
#> 2007-07-02 121.05 122.09 119.30 121.26 248715600  15.71039
#> 2007-07-03 122.00 127.40 121.50 127.17 290620400  16.47609
#> 2007-07-05 128.80 132.97 128.69 132.75 363262900  17.19903
#> 2007-07-06 133.13 133.34 130.40 132.30 218673700  17.14073
#> 2007-07-09 132.38 132.90 129.18 130.33 248955000  16.88550
#> 2007-07-10 128.88 134.50 128.81 132.35 313751900  17.14721
#> 2007-07-11 132.07 133.70 131.31 132.39 205443000  17.15239
#> 2007-07-12 133.85 134.24 132.39 134.07 176152200  17.37005
#> 2007-07-13 135.03 137.85 134.52 137.73 226901500  17.84424
#> 2007-07-16 138.39 139.98 137.50 138.10 234028200  17.89217
#> 2007-07-17 138.30 139.60 137.50 138.91 177489900  17.99712
#> 2007-07-18 138.19 138.44 136.04 138.12 189214200  17.89476
#> 2007-07-19 140.30 140.81 139.65 140.00 183222900  18.13834
#> 2007-07-20 141.65 144.18 140.00 143.75 291943400  18.62419
#> 2007-07-23 143.31 145.22 140.93 143.70 259122500  18.61771
#> 2007-07-24 138.88 141.00 134.15 134.89 448823200  17.47629
#> 2007-07-25 137.35 138.36 135.00 137.26 374045700  17.78334
#> 2007-07-26 145.91 148.50 136.96 146.00 546657300  18.91569
#> 2007-07-27 146.19 148.92 143.78 143.85 290274600  18.63714
#> 2007-07-30 144.33 145.45 139.57 141.43 276747100  18.32361
#> 2007-07-31 142.97 143.48 131.52 131.76 440598200  17.07077
#> 2007-08-01 133.64 135.38 127.77 135.00 437539200  17.49054
#> 2007-08-02 136.65 136.96 134.15 136.49 213161200  17.68358
#> 2007-08-03 135.26 135.95 131.50 131.85 169796900  17.08243
#> 2007-08-06 132.90 135.27 128.30 135.25 231292600  17.52293
#> 2007-08-07 134.94 137.24 132.63 135.03 237484100  17.49443
#> 2007-08-08 136.76 136.86 132.00 134.01 202024200  17.36228
#> 2007-08-09 131.11 133.00 125.09 126.39 281348900  16.37503
#> 2007-08-10 123.12 127.75 120.30 125.00 352687300  16.19494
#> 2007-08-13 128.32 129.35 126.50 127.79 188227900  16.55642
#> 2007-08-14 128.29 128.30 123.71 124.03 184751700  16.06927
#> 2007-08-15 122.74 124.86 119.65 119.90 248213000  15.53419
#> 2007-08-16 117.01 118.50 111.62 117.05 466672500  15.16494
#> 2007-08-17 122.01 123.50 119.82 122.06 298765600  15.81404
#> 2007-08-20 123.96 124.50 120.50 122.22 200829300  15.83477
#> 2007-08-21 122.21 128.96 121.00 127.57 325761800  16.52791
#> 2007-08-22 131.22 132.75 130.33 132.51 265441400  17.16794
#> 2007-08-23 133.09 133.34 129.76 131.07 216709500  16.98137
#> 2007-08-24 130.53 135.37 129.81 135.30 227958500  17.52941
#> 2007-08-27 133.39 134.66 132.10 132.25 176859900  17.13425
#> 2007-08-28 130.99 132.41 126.63 126.82 294841400  16.43074
#> 2007-08-29 129.88 134.18 129.54 134.08 291715200  17.37134
#>  [ reached getOption("max.print") -- omitted 2419 rows ]