Get all stocks in a stock index or stock exchange in tibble format

Get all stocks in a stock index or stock exchange in tibble format

tq_index(x, use_fallback = FALSE)

tq_exchange(x)

tq_index_options()

tq_exchange_options()

Arguments

x
A single character string, a character vector or tibble representing a single stock index or multiple stock indexes.

use_fallback
A boolean that can be used to return a fallback data set last downloaded when the package was updated. Useful if the website is down. Set to FALSE by default.

Value

Returns data in the form of a tibble object.

Details

tq_index() returns the stock symbol, company name, weight, and sector of every stock in an index. The source is www.ssga.com.

tq_index_options() returns a list of stock indexes you can choose from.

tq_exchange() returns the stock symbol, company, last sale price, market capitalization, sector and industry of every stock in an exchange. Three stock exchanges are available (AMEX, NASDAQ, and NYSE).

tq_exchange_options() returns a list of stock exchanges you can choose from. The options are AMEX, NASDAQ and NYSE.

See also

Examples

# Load libraries
library(tidyquant)

# Get the list of stock index options
tq_index_options()

#> [1] "DOW"       "DOWGLOBAL" "SP400"     "SP500"     "SP600"    

# Get all stock symbols in a stock index
if (FALSE) {
tq_index("DOW")
}

# Get the list of stock exchange options
tq_exchange_options()

#> [1] "AMEX"   "NASDAQ" "NYSE"  

# Get all stocks in a stock exchange
if (FALSE) {
tq_exchange("NYSE")
}