Get quantitative data in tibble format

Get quantitative data in tibble format

tq_get(x, get = "stock.prices", complete_cases = TRUE, ...)

tq_get_options()

tq_get_stock_index_options()

Arguments

x
A single character string, a character vector or tibble representing a single (or multiple) stock symbol, metal symbol, currency combination, FRED code, etc.

get
A character string representing the type of data to get for x. Options include:

"stock.prices": Get the open, high, low, close, volume and adjusted stock prices for a stock symbol from Yahoo Finance. Wrapper for quantmod::getSymbols().

"stock.prices.google": DISCONTINUED.

"stock.prices.japan": Get the open, high, low, close, volume and adjusted stock prices for a stock symbol from Yahoo Finance Japan. Wrapper for quantmod::getSymbols.yahooj().

"financials": DISCONTINUED.

"key.ratios": DISCONTINUED.

"key.stats": DISCONTINUED.

"dividends": Get the dividends for a stock symbol from Yahoo Finance. Wrapper for quantmod::getDividends().

"splits": Get the splits for a stock symbol from Yahoo Finance. Wrapper for quantmod::getSplits().

"economic.data": Get economic data from FRED. rapper for quantmod::getSymbols.FRED().

"metal.prices": Get the metal prices from Oanda. Wrapper for quantmod::getMetals().

"exchange.rates": Get exchange rates from Oanda. Wrapper for quantmod::getFX().

"quandl": Get data sets from Quandl. Wrapper for Quandl::Quandl(). See also quandl_api_key().

"quandl.datatable": Get data tables from Quandl. Wrapper for Quandl::Quandl.datatable(). See also quandl_api_key().

"alphavantager": Get data sets from Alpha Vantage. Wrapper for alphavantager::av_get(). See also av_api_key().

"rblpapi": Get data sets from Bloomberg. Wrapper for Rblpapi. See also Rblpapi::blpConnect() to connect to Bloomberg terminal (required). Use the argument rblpapi_fun to set the function such as "bdh" (default), "bds", or "bdp".

complete_cases
Removes symbols that return an NA value due to an error with the get call such as sending an incorrect symbol "XYZ" to get = "stock.prices". This is useful in scaling so user does not need to add an extra step to remove these rows. TRUE by default, and a warning message is generated for any rows removed.

...
Additional parameters passed to the "wrapped" function. Investigate underlying functions to see full list of arguments. Common optional parameters include:

from: Optional for various time series functions in quantmod / quandl packages. A character string representing a start date in YYYY-MM-DD format. No effect on "key.ratios", or "key.stats".

to: Optional for various time series functions in quantmod / quandl packages. A character string representing a end date in YYYY-MM-DD format. No effect on get = "key.ratios" or "key.stats".

Value

Returns data in the form of a tibble object.

Details

tq_get() is a consolidated function that gets data from various web sources. The function is a wrapper for several quantmod functions, Quandl functions, and also gets data from websources unavailable in other packages. The results are always returned as a tibble. The advantages are (1) only one function is needed for all data sources and (2) the function can be seemlessly used with the tidyverse: purrr, tidyr, and dplyr verbs.

tq_get_options() returns a list of valid get options you can choose from.

tq_get_stock_index_options() Is deprecated and will be removed in the next version. Please use tq_index_options() instead.

See also

Examples

# Load libraries
library(tidyquant)
library(dplyr)

# Get the list of `get` options
tq_get_options()

#>  [1] "stock.prices"        "stock.prices.google" "stock.prices.japan" 
#>  [4] "financials"          "key.ratios"          "dividends"          
#>  [7] "splits"              "economic.data"       "exchange.rates"     
#> [10] "metal.prices"        "quandl"              "quandl.datatable"   
#> [13] "alphavantager"       "rblpapi"            

# Get stock prices for a stock from Yahoo
aapl_stock_prices <- tq_get("AAPL")

# Get stock prices for multiple stocks
mult_stocks <- tq_get(c("FB", "AMZN"),
                      get  = "stock.prices",
                      from = "2016-01-01",
                      to   = "2017-01-01")

# Multiple gets
mult_gets <- tq_get("AAPL",
                    get = c("stock.prices", "dividends"),
                    from = "2016-01-01",
                    to   = "2017-01-01")