Mutates quantitative data

tq_mutate() adds new variables to an existing tibble; tq_transmute() returns only newly created columns and is typically used when periodicity changes

tq_mutate(data, select = NULL, mutate_fun, col_rename = NULL,
  ohlc_fun = NULL, ...)

tq_mutate_(data, select = NULL, mutate_fun, col_rename = NULL, ...)

tq_mutate_xy(data, x, y = NULL, mutate_fun, col_rename = NULL, ...)

tq_mutate_xy_(data, x, y = NULL, mutate_fun, col_rename = NULL, ...)

tq_mutate_fun_options()

tq_transmute(data, select = NULL, mutate_fun, col_rename = NULL,
  ohlc_fun = NULL, ...)

tq_transmute_(data, select = NULL, mutate_fun, col_rename = NULL, ...)

tq_transmute_xy(data, x, y = NULL, mutate_fun, col_rename = NULL, ...)

tq_transmute_xy_(data, x, y = NULL, mutate_fun, col_rename = NULL, ...)

tq_transmute_fun_options()

Arguments

data
A tibble (tidy data frame) of data typically from tq_get().

select
The columns to send to the mutation function.

mutate_fun
The mutation function from either the xts, quantmod, or TTR package. Execute tq_mutate_fun_options() to see the full list of options by package.

col_rename
A string or character vector containing names that can be used to quickly rename columns.

ohlc_fun
Deprecated. Use select.

...
Additional parameters passed to the appropriate mutatation function.

x, y
Parameters used with _xy that consist of column names of variables to be passed to the mutatation function (instead of OHLC functions).

Value

Returns mutated data in the form of a tibble object.

Details

tq_mutate and tq_transmute are very flexible wrappers for various xts, quantmod and TTR functions. The main advantage is the results are returned as a tibble and the function can be used with the tidyverse. tq_mutate is used when additional columns are added to the return data frame. tq_transmute works exactly like tq_mutate except it only returns the newly created columns. This is helpful when changing periodicity where the new columns would not have the same number of rows as the original tibble. select specifies the columns that get passed to the mutation function. Select works as a more flexible version of the OHLC extractor functions from quantmod where non-OHLC data works as well. When select is NULL, all columns are selected. In Example 1 below, close returns the "close" price and sends this to the mutate function, periodReturn. mutate_fun is the function that performs the work. In Example 1, this is periodReturn, which calculates the period returns. The ... are additional arguments passed to the mutate_fun. Think of the whole operation in Example 1 as the close price, obtained by select = close, being sent to the periodReturn function along with additional arguments defining how to perform the period return, which includes period = "daily" and type = "log". Example 4 shows how to apply a rolling regression. tq_mutate_xy and tq_transmute_xy are designed to enable working with mutatation functions that require two primary inputs (e.g. EVWMA, VWAP, etc). Example 2 shows this benefit in action: using the EVWMA function that uses volume to define the moving average period. tq_mutate_, tq_mutate_xy_, tq_transmute_, and tq_transmute_xy_ are setup for Non-Standard Evaluation (NSE). This enables programatically changing column names by modifying the text representations. Example 5 shows the difference in implementation. Note that character strings are being passed to the variables instead of unquoted variable names. See vignette("nse") for more information. tq_mutate_fun_options and tq_transmute_fun_options return a list of various financial functions that are compatible with tq_mutate and tq_transmute, respectively.

See also

tq_get()

Examples

# Load libraries
library(tidyquant)

##### Basic Functionality

fb_stock_prices  <- tq_get("FB",
                           get  = "stock.prices",
                           from = "2016-01-01",
                           to   = "2016-12-31")

# Example 1: Return logarithmic daily returns using periodReturn()
fb_stock_prices %>%
    tq_mutate(select = close, mutate_fun = periodReturn,
              period = "daily", type = "log")

#> # A tibble: 252 × 8
#>          date   open   high    low  close   volume adjusted daily.returns
#>        <date>  <dbl>  <dbl>  <dbl>  <dbl>    <dbl>    <dbl>         <dbl>
#> 1  2016-01-04 101.95 102.24  99.75 102.22 37912400   102.22   0.000000000
#> 2  2016-01-05 102.89 103.71 101.67 102.73 23258200   102.73   0.004976853
#> 3  2016-01-06 101.13 103.77 100.90 102.97 25096200   102.97   0.002333477
#> 4  2016-01-07 100.50 101.43  97.30  97.92 45172900    97.92  -0.050286895
#> 5  2016-01-08  99.88 100.50  97.03  97.33 35402300    97.33  -0.006043511
#> 6  2016-01-11  97.91  98.60  95.39  97.51 29932400    97.51   0.001847670
#> 7  2016-01-12  99.00  99.96  97.55  99.37 28395400    99.37   0.018895330
#> 8  2016-01-13 100.58 100.58  95.21  95.44 33410600    95.44  -0.040352489
#> 9  2016-01-14  95.85  98.87  92.45  98.37 48658600    98.37   0.030238111
#> 10 2016-01-15  93.98  96.38  93.54  94.97 45935600    94.97  -0.035174847
#> # ... with 242 more rows

# Example 2: Use tq_mutate_xy to use functions with two columns required
fb_stock_prices %>%
    tq_mutate_xy(x = close, y = volume, mutate_fun = EVWMA,
                 col_rename = "EVWMA")

#> # A tibble: 252 × 8
#>          date   open   high    low  close   volume adjusted EVWMA
#>        <date>  <dbl>  <dbl>  <dbl>  <dbl>    <dbl>    <dbl> <dbl>
#> 1  2016-01-04 101.95 102.24  99.75 102.22 37912400   102.22    NA
#> 2  2016-01-05 102.89 103.71 101.67 102.73 23258200   102.73    NA
#> 3  2016-01-06 101.13 103.77 100.90 102.97 25096200   102.97    NA
#> 4  2016-01-07 100.50 101.43  97.30  97.92 45172900    97.92    NA
#> 5  2016-01-08  99.88 100.50  97.03  97.33 35402300    97.33    NA
#> 6  2016-01-11  97.91  98.60  95.39  97.51 29932400    97.51    NA
#> 7  2016-01-12  99.00  99.96  97.55  99.37 28395400    99.37    NA
#> 8  2016-01-13 100.58 100.58  95.21  95.44 33410600    95.44    NA
#> 9  2016-01-14  95.85  98.87  92.45  98.37 48658600    98.37    NA
#> 10 2016-01-15  93.98  96.38  93.54  94.97 45935600    94.97 94.97
#> # ... with 242 more rows

# Example 3: Using tq_mutate to work with non-OHLC data
tq_get("DCOILWTICO", get = "economic.data") %>%
    tq_mutate(select = price, mutate_fun = lag.xts, k = 1, na.pad = TRUE)

#> # A tibble: 2,676 × 3
#>          date price lag.xts
#>        <date> <dbl>   <dbl>
#> 1  2007-01-01    NA      NA
#> 2  2007-01-02 60.77      NA
#> 3  2007-01-03 58.31   60.77
#> 4  2007-01-04 55.65   58.31
#> 5  2007-01-05 56.29   55.65
#> 6  2007-01-08 56.08   56.29
#> 7  2007-01-09 55.65   56.08
#> 8  2007-01-10 53.95   55.65
#> 9  2007-01-11 51.91   53.95
#> 10 2007-01-12 52.96   51.91
#> # ... with 2,666 more rows

# Example 4: Using tq_mutate to apply a rolling regression
fb_returns <- fb_stock_prices %>%
    tq_transmute(adjusted, periodReturn, period = "monthly", col_rename = "fb.returns")
xlk_returns <- tq_get("XLK", from = "2016-01-01", to = "2016-12-31") %>%
    tq_transmute(adjusted, periodReturn, period = "monthly", col_rename = "xlk.returns")
returns_combined <- left_join(fb_returns, xlk_returns, by = "date")
regr_fun <- function(data) {
    coef(lm(fb.returns ~ xlk.returns, data = as_data_frame(data)))
}
returns_combined %>%
    tq_mutate(mutate_fun = rollapply,
              width      = 6,
              FUN        = regr_fun,
              by.column  = FALSE,
              col_rename = c("coef.0", "coef.1"))

#> # A tibble: 12 × 5
#>          date  fb.returns  xlk.returns       coef.0    coef.1
#>        <date>       <dbl>        <dbl>        <dbl>     <dbl>
#> 1  2016-01-29  0.09773036 -0.024367147           NA        NA
#> 2  2016-02-29 -0.04714376 -0.006547059           NA        NA
#> 3  2016-03-31  0.06715301  0.088216899           NA        NA
#> 4  2016-04-29  0.03049960 -0.050270511           NA        NA
#> 5  2016-05-31  0.01046093  0.048896212           NA        NA
#> 6  2016-06-30 -0.03812810 -0.013777982  0.018763741 0.1895498
#> 7  2016-07-29  0.08452925  0.071016877  0.003548089 0.6258971
#> 8  2016-08-31  0.01758916  0.011625439  0.015421792 0.5110441
#> 9  2016-09-30  0.01704726  0.021017965  0.013648536 0.4531443
#> 10 2016-10-31  0.02120528 -0.007534533 -0.001427809 0.9240092
#> 11 2016-11-30 -0.09596157  0.001687075 -0.018894284 1.4237784
#> 12 2016-12-30 -0.02845799  0.022598376 -0.025422028 1.3992350

# Example 5: Non-standard evaluation:
# Programming with tq_mutate_() and tq_mutate_xy_()
col_name <- "adjusted"
mutate <- c("MACD", "SMA")
tq_mutate_xy_(fb_stock_prices, x = col_name, mutate_fun = mutate[[1]])

#> # A tibble: 252 × 9
#>          date   open   high    low  close   volume adjusted  macd signal
#>        <date>  <dbl>  <dbl>  <dbl>  <dbl>    <dbl>    <dbl> <dbl>  <dbl>
#> 1  2016-01-04 101.95 102.24  99.75 102.22 37912400   102.22    NA     NA
#> 2  2016-01-05 102.89 103.71 101.67 102.73 23258200   102.73    NA     NA
#> 3  2016-01-06 101.13 103.77 100.90 102.97 25096200   102.97    NA     NA
#> 4  2016-01-07 100.50 101.43  97.30  97.92 45172900    97.92    NA     NA
#> 5  2016-01-08  99.88 100.50  97.03  97.33 35402300    97.33    NA     NA
#> 6  2016-01-11  97.91  98.60  95.39  97.51 29932400    97.51    NA     NA
#> 7  2016-01-12  99.00  99.96  97.55  99.37 28395400    99.37    NA     NA
#> 8  2016-01-13 100.58 100.58  95.21  95.44 33410600    95.44    NA     NA
#> 9  2016-01-14  95.85  98.87  92.45  98.37 48658600    98.37    NA     NA
#> 10 2016-01-15  93.98  96.38  93.54  94.97 45935600    94.97    NA     NA
#> # ... with 242 more rows